use one test for all fs
